pyproject.toml
[tool.poetry]
name = "quich"
version = "0.1.0"
description = ""
authors = ["falcolnic <mykh.yano@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.115.2"
pydantic =  {extras = ["email"], version = "2.9.2"}

sqlalchemy = "2.0.35"
uvicorn = "^0.32.0"
voyageai = "0.2.3"
anthropic = "^0.45.2"
instructor = "1.6.3"
fasthx = "2.0.1"
pydantic-settings = "^2.7.1"
pyjwt = "^2.10.1"
python-multipart = "^0.0.20"
sqlite-vec = "^0.1.6"
youtube-transcript-api = "0.6.2"
pytubefix = "^8.12.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg2-binary = "2.9.9"

[tool.poetry.group.dev.dependencies]
black = "24.4.2"
pre-commit = "3.8.0"
isort = "5.13.2"
ruff = "0.5.5"
psycopg2-binary = "2.9.9"
pytest = "8.3.3"
alembic = "1.13.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
